webvtt/parsing/file-parsing/tests/regions-edge-case.html

<!doctype html>
<!-- DO NOT EDIT! This file and support/regions-edge-case.vtt are generated. -->
<!-- See /webvtt/parsing/file-parsing/README.md -->
<meta charset=utf-8>
<title>WebVTT parser test: multiple regions edge cases</title>
<link rel="help" href="https://w3c.github.io/webvtt/#collect-webvtt-region-settings">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
var t = async_test('regions, lines');
t.step(function(){
    var video = document.createElement('video');
    var track = document.createElement('track');
    assert_true('src' in track, 'track element not supported');
    track.src = 'support/regions-edge-case.vtt';
    track['default'] = true;
    track.kind = 'subtitles';
    track.onload = this.step_func(trackLoaded);
    track.onerror = this.step_func(trackError);
    video.appendChild(track);
    document.body.appendChild(video);
});

function trackLoaded(event) {
    var track = event.target;
    var video = track.parentNode;
    var cues = video.textTracks[0].cues;
    {
assert_equals(cues.length, 4);

var regions = Array.from(cues).map(function(cue) {
    return cue.region;
});

var valid_lines = [
    1,
    2,
    3,
    4
];
valid_lines.forEach(function(valid, index) {
    assert_equals(regions[index].lines, valid, 'Failed with region ' + index);
});
    }
    this.done();
}

function trackError(e) {
    assert_unreached('got unexpected error event');
}
</script>
